updated gemspec & install

+++ install.sh
@@ -1,4 +1,4 @@
-rm pkg/devOn-1.0.1.gem
+[ -f pkg/devOn-1.0.1.gem ] && rm pkg/devOn-1.0.1.gem
 gem uninstall devOn
 rake build
 gem install pkg/devOn-1.0.1.gem
